refactor(protocoltest): split duo config by role and move the spec to duo_spec.go
Follow-up to review discussion on DuoEnvConfig.TB2HTTPTimeouts:

- DuoEnvConfig now mirrors the child-side role split: Upstream{Stream}
  configures tb1, Gateway{HTTPTimeouts} configures tb2, and top-level
  fields (ChildLog, BootTimeout) are the ones that genuinely apply to the
  environment as a whole. Instance-specific parameters get a structural
  home instead of a name-prefixed field (TB2HTTPTimeouts), and Stream's
  real ownership (tb1-only) is now explicit too.
- The boot contract (duoEnvSpec, duoRole, duoInstanceSpec, DuoStreamShape,
  validate/encode/decode) moves out of duo_serve.go into its own
  duo_spec.go, matching its dedicated duo_spec_test.go.
